Backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
PyJWT==2.8.0
pandas==2.1.0
scikit-learn==1.3.0
xgboost==1.7.6
pymongo==4.5.0
